Rename the aligner module to make it generic
