+++ scripts/run_two_video_sync_demo.py
@@ -15,8 +15,8 @@ import numpy as np
 from mne.datasets import sample
 from qtpy.QtWidgets import QApplication
 
-from videomeg_browser.raw_media_aligner import TimestampAligner
 from videomeg_browser.synced_raw_media_browser import browse_raw_with_video
+from videomeg_browser.timestamp_aligner import TimestampAligner
 from videomeg_browser.video import VideoFileHelsinkiVideoMEG
