+++ scripts/run_video_sync_demo.py
@@ -13,8 +13,8 @@ from numpy.typing import NDArray
 from qtpy.QtWidgets import QApplication
 
 from videomeg_browser.comp_tstamps import comp_tstamps
-from videomeg_browser.raw_media_aligner import TimestampAligner
 from videomeg_browser.synced_raw_media_browser import browse_raw_with_video
+from videomeg_browser.timestamp_aligner import TimestampAligner
 from videomeg_browser.video import VideoFileHelsinkiVideoMEG
