+++ scripts/run_video_sync_demo_2.py
@@ -11,9 +11,9 @@ import scipy
 from mne.datasets import sample
 from qtpy.QtWidgets import QApplication
 
-from videomeg_browser.raw_media_aligner import TimestampAligner
 from videomeg_browser.synced_raw_media_browser import browse_raw_with_video
 from videomeg_browser.test_utils import create_fake_video_with_markers
+from videomeg_browser.timestamp_aligner import TimestampAligner
 
 
 def _create_binary_stimulus_vector(
