+++ src/videomeg_browser/timestamp_aligner.py
